ther was a bug in your code
We found a bug in the code, which caused the wrong display in RootElement when the radio groups wer recreated. The previous states were "saved" as the cell-s are reused by the UITableView and you missed to remove the detail text ot this cell-s if no selected item in the current radiogroup.

We made some changes in StyledElement too where was wrong display of row height when the content was higher than the cell and the rowheight, And added a sample how to recalculate the size of the table to display after added all the rows to it.